better input widgets; move route to project.controller; rename as settings.mustache
