--- controller/newProject/index.js
@@ -1,9 +1,0 @@
-var express = require('express');
-var controller = require('./newProject.controller');
-
-var router = express.Router();
-
-router.get('/:projectName', controller.validator , controller.newProject);
-
-
-module.exports = router;
