+++ controller/project/index.js
@@ -4,8 +4,9 @@ var controller = require('./project.controller');
 var router = express.Router();
 
 //
+router.get('/new', controller.newProject);
 router.get('/:projectName', controller.validator , controller.project);
+router.get('/:projectName/settings', controller.validator , controller.settings);
 router.get('/json/:projectName', controller.validator , controller.api_project);
-
 
 module.exports = router;
